requirements.txt
pandas
numpy
seaborn
matplotlib
notebook
dvc
pyYAML
types-pyYAML
mlflow
tensorflow
python-box
ensure
tqdm
joblib
scipy
Flask
Flask-Cors
gdown
-e .